requirements.txt
deepl>=1.7
discord-py-interactions>=4.2
enhanced>=4.0
